@@ -1,4 +1,4 @@
- /*
+/*
  * Copyright 2014 samuelcampos.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -33,20 +33,14 @@ public abstract class AbstractStorageDeviceDetector {
   public static synchronized AbstractStorageDeviceDetector getInstance() {
     if (instance == null) {
       switch (OSUtils.getOsType()) {
-        case WINDOWS:
+        case WINDOWS ->
           instance = new WindowsStorageDeviceDetector();
-          break;
-
-        case LINUX:
+        case LINUX ->
           instance = new LinuxStorageDeviceDetector();
-          break;
-
-        case MAC_OS:
+        case MAC_OS ->
           instance = new OSXStorageDeviceDetector();
-          break;
       }
     }
-
     return instance;
   }
 
@@ -61,7 +55,6 @@ public abstract class AbstractStorageDeviceDetector {
 
   static Optional<USBStorageDevice> getUSBDevice(final String rootPath, final String deviceName, final String device, final String uuid) {
     final File root = new File(rootPath);
-
     if (!root.isDirectory()) {
       LOGGER.info(tl("usb_detector_invalid_root_found", root));
       return Optional.empty();
@@ -73,7 +66,6 @@ public abstract class AbstractStorageDeviceDetector {
       LOGGER.info(tl("usb_detector_could_not_add_device", e.getMessage()));
       LOGGER.exception(e);
     }
-
     return Optional.empty();
   }
 
