@@ -55,18 +55,11 @@ public class LinuxStorageDeviceDetector extends AbstractStorageDeviceDetector {
 
         if (parts.length > 1) {
           switch (parts[0].trim()) {
-            case INFO_BUS:
-              disk.setUSB(INFO_USB.equals(parts[1].trim()));
-              break;
-            case INFO_NAME:
-              disk.setName(parts[1].trim());
-              break;
-            case INFO_UUID:
-              disk.setUuid(parts[1].trim());
-              break;
+            case INFO_BUS -> disk.setUSB(INFO_USB.equals(parts[1].trim()));
+            case INFO_NAME -> disk.setName(parts[1].trim());
+            case INFO_UUID -> disk.setUuid(parts[1].trim());
           }
         }
-
       });
 
     } catch (final IOException e) {
