@@ -122,21 +122,10 @@ public class OSXStorageDeviceDetector extends AbstractStorageDeviceDetector {
 
         if (parts.length > 1) {
           switch (parts[0].trim()) {
-            case INFO_MOUNTPOINT:
-              disk.setMountPoint(parts[1].trim());
-              break;
-
-            case INFO_PROTOCOL:
-              disk.setUSB(INFO_USB.equals(parts[1].trim()));
-              break;
-
-            case INFO_NAME:
-              disk.setName(parts[1].trim());
-              break;
-
-            case INFO_UUID:
-              disk.setUuid(parts[1].trim());
-              break;
+            case INFO_MOUNTPOINT -> disk.setMountPoint(parts[1].trim());
+            case INFO_PROTOCOL -> disk.setUSB(INFO_USB.equals(parts[1].trim()));
+            case INFO_NAME -> disk.setName(parts[1].trim());
+            case INFO_UUID -> disk.setUuid(parts[1].trim());
           }
         }
 
