@@ -29,20 +29,14 @@ public abstract class AbstractStorageDeviceUnmounter {
   public static synchronized AbstractStorageDeviceUnmounter getInstance() {
     if (instance == null) {
       switch (OSUtils.getOsType()) {
-        case WINDOWS:
+        case WINDOWS ->
           instance = new WindowsStorageDeviceUnmounter();
-          break;
-
-        case LINUX:
+        case LINUX ->
           instance = new LinuxStorageDeviceUnmounter();
-          break;
-
-        case MAC_OS:
+        case MAC_OS ->
           instance = new OSXStorageDeviceUnmounter();
-          break;
       }
     }
-
     return instance;
   }
 
