@@ -17,6 +17,7 @@
 package optic_fusion1.antimalware;
 
 import java.io.File;
+import optic_fusion1.antimalware.check.CacheContainer;
 import optic_fusion1.antimalware.check.CheckManager;
 import optic_fusion1.antimalware.check.CheckRegistery;
 import optic_fusion1.antimalware.database.Database;
@@ -43,6 +44,7 @@ public class AntiMalware extends Thread {
   private MetricsManager metrics;
   private Scanner scanner;
   private ServerHandler serverHandler;
+  private CacheContainer cache;
 
   public AntiMalware(String[] args) {
     setName("AntiMalware/Main");
@@ -62,6 +64,7 @@ public class AntiMalware extends Thread {
   private void init() {
     LOGGER.info(I18n.tl("initializing_start"));
     LOGGER.info(I18n.tl("github_issues"));
+    cache = new CacheContainer(DATABASE);
     notifHandler = new NotificationHandler(parser.getNotificationType());
     if (parser.shouldUseMetrics()) {
       (metrics = new MetricsManager(this)).registerMetrics(new JVMMetric(), new MemoryMetric()).start();
@@ -116,11 +119,6 @@ public class AntiMalware extends Thread {
     return DATABASE;
   }
 
-  // TODO: Replace this with a better method eventually
-  public static Database getDB() {
-    return DATABASE;
-  }
-
   public static AntiMalware getInstance() {
     return instance;
   }
@@ -135,6 +133,10 @@ public class AntiMalware extends Thread {
 
   public ServerHandler getServerHandler() {
     return serverHandler;
+  }
+
+  public CacheContainer getCache() {
+    return cache;
   }
 
 }
