@@ -28,7 +28,8 @@ public abstract class BaseCheck {
   private String sourceFilePath = "";
   private int line = -1;
 
-  public abstract List<CheckResult> process(ClassNode classNode, Path rootFolder, Path zipFile, CacheContainer cache);
+  public abstract List<CheckResult> process(ClassNode classNode, Path rootFolder, Path zipFile,
+          CacheContainer cache);
 
   public boolean isPlugin(Path rootDir) {
     return Files.exists(rootDir.resolve("plugin.yml")) || Files.exists(rootDir.resolve("bungee.yml"));
