@@ -46,7 +46,14 @@ public class CacheContainer {
   private HashSet<String> blacklistedAuthors;
   private HashSet<String> blacklistedWebsites;
   private HashSet<String> blacklistedCommands;
-
+  private HashSet<String> blacklistedUrls;
+  private HashSet<String> blacklistedExec;
+  private HashSet<String> blacklistedWrite;
+  private HashSet<String> blacklistedChecksums;
+  private HashSet<String> blacklistedFileNames;
+  private HashSet<String> blacklistedFilePaths;
+  private HashSet<String> whitelistedChecksums;
+  
   public CacheContainer(Database database) {
     try {
       blacklistedClasspaths = database.getBlacklistedClasspaths();
@@ -54,6 +61,13 @@ public class CacheContainer {
       blacklistedAuthors = database.getBlacklistedAuthors();
       blacklistedWebsites = database.getBlacklistedWebsites();
       blacklistedCommands = database.getBlacklistedCommands();
+      blacklistedUrls = database.getBlacklistedUrls();
+      blacklistedExec = database.getBlacklistedExec();
+      blacklistedWrite = database.getBlacklistedWrite();
+      blacklistedChecksums = database.getBlacklistedChecksums();
+      blacklistedFileNames = database.getBlacklistedFileNames();
+      blacklistedFilePaths = database.getBlacklistedFilePaths();
+      whitelistedChecksums = database.getWhitelistedChecksums();
     } catch (SQLException ex) {
       LOGGER.exception(ex);
     }
@@ -142,15 +156,43 @@ public class CacheContainer {
   public boolean containsBlacklistedClassPath(String classPath) {
     return blacklistedClasspaths.contains(classPath);
   }
-  
-  public boolean containsBlacklistedString(String string){
+
+  public boolean containsBlacklistedString(String string) {
     return blacklistedStrings.contains(string);
   }
 
-  public boolean containsBlacklistedCommand(String command){
+  public boolean containsBlacklistedCommand(String command) {
     return blacklistedCommands.contains(command);
   }
 
+  public boolean containsBlacklistedUrl(String url) {
+    return blacklistedUrls.contains(url);
+  }
+
+  public boolean containsBlacklistedExec(String exec) {
+    return blacklistedExec.contains(exec);
+  }
+
+  public boolean containsBlacklistedWrite(String file) {
+    return blacklistedWrite.contains(file);
+  }
+  
+  public boolean containsBlacklistedChecksum(String checksum){
+    return blacklistedChecksums.contains(checksum);
+  }
+  
+  public boolean containsBlacklistedFileName(String fileName){
+    return blacklistedFileNames.contains(fileName);
+  }
+
+  public boolean containsBlacklistedFilePath(String filePath){
+    return blacklistedFilePaths.contains(filePath);
+  }
+  
+  public boolean containsWhitelistedChecksum(String checksum){
+    return whitelistedChecksums.contains(checksum);
+  }
+  
   private static class CacheMap<T> {
 
     private final Function<? super Path, ? extends Map<Path, T>> nestedCreator = p -> new HashMap<>();
