@@ -47,7 +47,7 @@ public class CheckManager {
     }
     try {
       String checksum = DigestUtils.sha1Hex(new FileInputStream(file));
-      boolean blacklisted = AntiMalware.getInstance().getDatabase().isChecksumBlacklisted(checksum.toUpperCase());
+      boolean blacklisted = AntiMalware.getInstance().getCache().containsBlacklistedChecksum(checksum.toUpperCase());
       CHECKSUM_CACHE.put(file, blacklisted);
       return blacklisted;
     } catch (FileNotFoundException ex) {
