@@ -34,7 +34,7 @@ public class CheckResult {
     this(platform, type, family, variant, null, null, -1);
   }
 
-  public CheckResult(String platform, String type, String family, String variant, String sourceFilePath,
+  public CheckResult(String platform, String type, String family, String variant, String sourceFilePath, 
           String classNodePath, int line) {
     this.platform = platform;
     this.type = type;
