@@ -72,7 +72,7 @@ public class Scanner extends Thread {
     setName("AntiMalware/Scanner");
     this.antiMalware = antiMalware;
     database = antiMalware.getDatabase();
-    cache = new CacheContainer(database);
+    cache = antiMalware.getCache();
     checkManager = antiMalware.getCheckManager();
     commandLineParser = antiMalware.getCommandLineParser();
     if (commandLineParser.shouldScanZippedFiles()) {
@@ -168,7 +168,7 @@ public class Scanner extends Thread {
       return;
     }
     String fileName = file.getFileName().toString();
-    if (database.isFileNameBlacklisted(fileName)) {
+    if (cache.containsBlacklistedFileName(fileName)) {
       try {
         CheckResult result = database.getCheckResultForFileName(fileName);
         if (result != null) {
@@ -179,7 +179,7 @@ public class Scanner extends Thread {
         LOGGER.exception(ex);
       }
     }
-    if (database.isFilePathBlacklisted(file.toString())) {
+    if (cache.containsBlacklistedFilePath(file.toString())) {
       try {
         CheckResult result = database.getCheckResultForFilePath(file.toString());
         if (result != null) {
@@ -201,14 +201,19 @@ public class Scanner extends Thread {
       LOGGER.exception(e);
       return;
     }
-    try {
-      CheckResult checkResult = database.getCheckResultForChecksum(cache.fetchSHA1(file, file));
-      if (checkResult != null) {
-        sendNotification(file, checkResult);
-        return;
+    String checksum = cache.fetchSHA1(file, file).toUpperCase();
+    if (cache.containsBlacklistedChecksum(checksum)) {
+      try {
+        CheckResult checkResult = database.getCheckResultForChecksum(checksum);
+        if (checkResult != null) {
+          sendNotification(file, checkResult);
+          return;
+        }
+        // TODO: Clean message up and make it translatable
+        LOGGER.alert(file + " is blacklisted by doesn't have a check result in the database, this should be reported to Optic_Fusion1");
+      } catch (SQLException ex) {
+        LOGGER.exception(ex);
       }
-    } catch (SQLException ex) {
-      Logger.getLogger(Scanner.class.getName()).log(Level.SEVERE, null, ex);
     }
     try ( FileSystem fs = Utils.fileSystemForZip(file)) {
       if (fs == null) {
@@ -337,7 +342,7 @@ public class Scanner extends Thread {
   }
 
   private WhitelistResult isChecksumWhitelisted(String checksum) {
-    return antiMalware.getDatabase().isChecksumWhitelisted(checksum) ? WhitelistResult.WHITELISTED
+    return antiMalware.getCache().containsWhitelistedChecksum(checksum) ? WhitelistResult.WHITELISTED
             : WhitelistResult.NOT_WHITELISTED;
   }
 
