@@ -73,10 +73,7 @@ public class PaperServerHandler extends ServerHandler {
     serverJar = new File(ppsm.getPaperPatchedJarFile());
     setSecurityManager(oldSM);
     this.version = ReflectionUtils.getVersion(serverJar);
-    if (version.equals(I18n.tl("minecraft_server_version_not_found"))) { //what?
-      return false;
-    }
-    return true;
+    return !version.equals(I18n.tl("minecraft_server_version_not_found"));
   }
 
   @Override
