@@ -39,10 +39,7 @@ public class SpigotServerHandler extends ServerHandler {
       return false;
     }
     this.version = ReflectionUtils.getVersion(getServerFile());
-    if (version.equals(I18n.tl("minecraft_server_version_not_found"))) {
-      return false;
-    }
-    return true;
+    return !version.equals(I18n.tl("minecraft_server_version_not_found"));
   }
 
   @Override
