@@ -24,7 +24,7 @@ import java.security.Permission;
 import optic_fusion1.antimalware.AntiMalware;
 import static optic_fusion1.antimalware.AntiMalware.LOGGER;
 import optic_fusion1.antimalware.CommandLineParser;
-import optic_fusion1.antimalware.database.Database;
+import optic_fusion1.antimalware.check.CacheContainer;
 import optic_fusion1.antimalware.exceptions.FormattedSecurityException;
 import optic_fusion1.antimalware.servers.runtimeprotect.callerinfo.CallerInfo;
 import optic_fusion1.antimalware.servers.runtimeprotect.callerinfo.RuntimeUtils;
@@ -32,12 +32,12 @@ import static optic_fusion1.antimalware.utils.I18n.tl;
 
 public class AMSecurityManager extends SecurityManager {
 
-  private Database database;
+  private CacheContainer cache;
   private CommandLineParser clp;
   private AntiMalware antiMalware;
 
   public AMSecurityManager(AntiMalware antiMalware) {
-    this.database = antiMalware.getDatabase();
+    cache = antiMalware.getCache();
     this.antiMalware = antiMalware;
     clp = antiMalware.getCommandLineParser();
   }
@@ -130,7 +130,7 @@ public class AMSecurityManager extends SecurityManager {
     CallerInfo callerInfo = RuntimeUtils.getCallerInfo();
     if (callerInfo != null) {
       File pluginJar = callerInfo.getPlugin().getJar();
-      if (database.isURLBlacklisted(host)) {
+      if (cache.containsBlacklistedUrl(host)) {
         throw new SecurityException(tl("security_manager_mal_connection", pluginJar, host, port));
       }
     }
@@ -160,7 +160,7 @@ public class AMSecurityManager extends SecurityManager {
     CallerInfo callerInfo = RuntimeUtils.getCallerInfo();
     if (callerInfo != null) {
       File pluginJar = callerInfo.getPlugin().getJar();
-      if (database.isWriteBlacklisted(file)) {
+      if (cache.containsBlacklistedWrite(file)) {
         throw new FormattedSecurityException(tl("security_manager_write_mal_file", pluginJar, file));
       }
       if (file.contains("AntiMalware")) {
@@ -220,7 +220,7 @@ public class AMSecurityManager extends SecurityManager {
     CallerInfo callerInfo = RuntimeUtils.getCallerInfo();
     if (callerInfo != null) {
       File pluginJar = callerInfo.getPlugin().getJar();
-      if (database.isExecBlacklisted(cmd)) {
+      if (cache.containsBlacklistedExec(cmd)) {
         throw new FormattedSecurityException(tl("security_manager_ran_exec", pluginJar, cmd));
       }
     }
