@@ -23,6 +23,7 @@ import net.bytebuddy.implementation.bind.annotation.SuperCall;
 import net.bytebuddy.implementation.bind.annotation.This;
 import optic_fusion1.antimalware.AntiMalware;
 import static optic_fusion1.antimalware.AntiMalware.LOGGER;
+import optic_fusion1.antimalware.check.CacheContainer;
 import optic_fusion1.antimalware.exceptions.FormattedSecurityException;
 import optic_fusion1.antimalware.servers.runtimeprotect.callerinfo.CallerInfo;
 import optic_fusion1.antimalware.servers.runtimeprotect.callerinfo.RuntimeUtils;
@@ -42,7 +43,11 @@ public final class AsyncPlayerChatEventTransformer {
       try {
         String message = (String) object.getClass().getMethod("getMessage").invoke(object);
         String player = getPlayerName(object);
-        if (AntiMalware.getInstance().getDatabase().isStringBlacklisted(message)) {
+        // TODO: Improve this check
+        CacheContainer cache = AntiMalware.getInstance().getCache();
+        if (cache.containsBlacklistedCommand(message) || cache.containsBlacklistedString(message)
+                || cache.containsBlacklistedString(message.split(" ")[0])
+                || cache.containsBlacklistedCommand(message.split(" ")[0])) {
           CallerInfo callerInfo = RuntimeUtils.getCallerInfo();
           if (callerInfo != null) {
             throw new FormattedSecurityException(tl("transformer_blacklisted_word",
